requirements.txt
discord.py
python-dotenv
pyyaml
rich
typing-extensions
aiohttp
asyncio
supabase